Add files via upload
6/2 AT HOME
Worked on a lot, did a lot of textures (stove, table, stove animation, frying pan, trash, etc, made them myself). Made a new screen for the cooking and stove, etc. Added a new object for kitchen appliances. Added hovering text when you hover over something. Added a place you can get the food but have not added the spawning of the food yet. Added a back button that works and goes back. Added turning on and off the stove and also fixed the customer stuff. Fixed some bugs.

Basically did all the busy work and things that take a long time at home, such as new screens, graphics, some of the functional parts, but mostly graphics animations and new screens and moving between the screens. Also set up a lot of conditions and got a better idea and plan of what I'm doing. Also fixed adding a bunch of things.