Add files via upload
Most productive update yet. Added a trash can that works to get rid of stuff on the screen. Made it so things on the screen stay when u enter another screen. Added a cooking feature which cooks your food based off how long it is on the stove. Added a feature that allows you to write click and pulls up a menu (kind of like when you right click your desktop background to pull up that menu) and that allows you to serve the food. If it is not possible to serve the food, it will flash red (proudest thing I've done yet).