Add files via upload
6/8 AND 6/9 
AT HOME
Got a lot done. Finished a lot of the sprites for the knife, bread, cutting board, salt, lettuce, chopped lettuce, carrot, chopped carrot. cooked steak, salads, served steak. the plate, the bowl, etc.
Implemented the feature for plating and combining the food to serve (finally done) and also implemented the salad. I have only made caesar salad so far and fruit salad is the next goal. Added a salt shaker too but that only works for salads.
Scrapped cutting the steak as the textures might be too much to make. Added the cutting system. Fixed a whole new system for food names. Created a new system to use plates and such.